+++ packages/expo-dev-menu/CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Fix compatibility with react-native 0.66. ([#15914](https://github.com/expo/expo/pull/15914) by [@kudo](https://github.com/kudo))
 - Fix Android crash when using Hermes on react-native 0.67. ([#16099](https://github.com/expo/expo/pull/16099) by [@kudo](https://github.com/kudo))
+- Fix backwards compatibility with AppDelegate in existing projects.
 
 ### 💡 Others